+++ flow/types/auth-user.js
@@ -1,1 +1,0 @@
-export type AuthUserType = {user: {id: number, roles: Array<{name: string}>}};