+++ src/lib/config.js
@@ -1,0 +1,24 @@
+// @flow
+
+let cfg = {
+  ignoredStackTraceLines: [],
+  captureUncaught: false,
+  captureUnhandled: false,
+  heretic: {},
+  db: {},
+  cache: {},
+  pugOptions: {
+    basedir: '',
+  },
+  email: {
+    devEmail: null,
+    guardedEnvironments: [],
+  },
+  pathToServerErrorTemplate: null,
+  enableNewRelic: false,
+};
+const cache = {};
+const heretic = {};
+const db = {};
+
+module.exports = {cfg, cache, db, heretic};